requirements.txt
streamlit>=1.28.0
pandas>=2.1.0
numpy>=1.24.0
python-dotenv>=1.0.0
python-binance==1.0.19
yfinance==0.2.28
requests==2.31.0
scikit-learn>=1.3.0
xgboost>=1.7.0
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0
joblib==1.3.2
tqdm==4.66.1